requirements.txt
streamlit==1.31.0
pandas==2.2.0
openpyxl==3.1.2
geopy==2.4.1
requests==2.31.0
python-dotenv